move resource view tests to their own file
