+++ ckanext/unhcr/tests/pt/controllers/test_dataset.py
@@ -2,10 +2,8 @@
 
 import mock
 import pytest
-from sqlalchemy import select, and_
 import ckan.model as model
 from ckan.plugins import toolkit
-from ckan.tests import helpers as core_helpers
 from ckantoolkit.tests import factories as core_factories
 from ckanext.unhcr.models import AccessRequest
 from ckanext.unhcr.tests import factories, mocks
@@ -63,21 +61,6 @@ class TestExtendedPackageController(object):
 
     def make_dataset_copy_request(self, app, dataset_id=None, user=None, **kwargs):
         url = '/dataset/copy/%s' % dataset_id
-        env = {'REMOTE_USER': user.encode('ascii')} if user else {}
-        resp = app.get(url=url, extra_environ=env, **kwargs)
-        return resp
-
-    def make_resource_copy_request(self, app, dataset_id=None, resource_id=None, user=None, **kwargs):
-        url = '/dataset/%s/resource_copy/%s' % (dataset_id, resource_id)
-        env = {'REMOTE_USER': user.encode('ascii')} if user else {}
-        resp = app.get(url=url, extra_environ=env, **kwargs)
-        return resp
-
-    def make_resource_download_request(self, app, dataset_id, resource_id, user=None, **kwargs):
-        url = '/dataset/{dataset}/resource/{resource}/download'.format(
-            dataset=dataset_id,
-            resource=resource_id,
-        )
         env = {'REMOTE_USER': user.encode('ascii')} if user else {}
         resp = app.get(url=url, extra_environ=env, **kwargs)
         return resp
@@ -153,6 +136,7 @@ class TestExtendedPackageController(object):
             status=200,
         )
 
+
     # Dataset Copy
 
     def test_dataset_copy(self, app):
@@ -201,152 +185,6 @@ class TestExtendedPackageController(object):
             status=404
         )
 
-    # Resource Upload
-
-    def test_edit_resource_works(self, app):
-        url = toolkit.url_for(
-            'resource.edit',
-            id=self.dataset1['id'],
-            resource_id=self.resource1['id']
-        )
-        env = {'REMOTE_USER': self.sysadmin['name'].encode('ascii')}
-
-        # Mock a resource edit payload
-        data = {
-            'id': self.resource1['id'],
-            'name': self.resource1['name'],
-            'type': self.resource1['type'],
-            'description': 'updated',
-            'format': self.resource1['format'],
-            'file_type': self.resource1['file_type'],
-            'date_range_start': self.resource1['date_range_start'],
-            'date_range_end': self.resource1['date_range_end'],
-            'version': self.resource1['version'],
-            'process_status': self.resource1['process_status'],
-            'identifiability': self.resource1['identifiability'],
-
-            'url': 'test.txt',
-            'save': ''
-
-        }
-
-        resp = app.post(url, data=data, extra_environ=env, status=200)
-
-        assert 'The form contains invalid entries:' not in resp.body
-
-    def test_edit_resource_must_provide_upload(self, app):
-        url = toolkit.url_for(
-            'resource.edit',
-            id=self.dataset1['id'],
-            resource_id=self.resource1['id']
-        )
-        env = {'REMOTE_USER': self.sysadmin['name'].encode('ascii')}
-
-        # Mock a resource edit payload
-        data = {
-            'id': self.resource1['id'],
-            'name': self.resource1['name'],
-            'type': self.resource1['type'],
-            'description': 'updated',
-            'format': self.resource1['format'],
-            'file_type': self.resource1['file_type'],
-            'date_range_start': self.resource1['date_range_start'],
-            'date_range_end': self.resource1['date_range_end'],
-            'version': self.resource1['version'],
-            'process_status': self.resource1['process_status'],
-            'identifiability': self.resource1['identifiability'],
-
-            'url': '',
-            'clear_upload': 'true',
-            'save': ''
-
-        }
-
-        resp = app.post(url, data=data, extra_environ=env, status=200)
-
-        assert 'The form contains invalid entries:' in resp.body
-        assert 'All data resources require an uploaded file' in resp.body
-
-    # Resource Copy
-
-    def test_resource_copy(self, app):
-        resp = self.make_resource_copy_request(
-            app, dataset_id='dataset1', resource_id=self.resource1['id'], user='user1',
-            status=200
-        )
-        assert 'action="/dataset/dataset1/resource/new"' in resp.body
-        assert 'resource1 (copy)' in resp.body
-        assert 'anonymized_public' in resp.body
-        assert 'Add'in resp.body
-
-    def test_resource_copy_no_access(self, app):
-        resp = self.make_resource_copy_request(
-            app, dataset_id='dataset1', resource_id=self.resource1['id'], user='user2',
-            status=403
-        )
-
-    def test_resource_copy_bad_resource(self, app):
-        resp = self.make_resource_copy_request(
-            app, dataset_id='dataset1', resource_id='bad', user='user1',
-            status=404
-        )
-
-    # Resource Download
-
-    def test_resource_download_anonymous(self, app):
-        resp = self.make_resource_download_request(
-            app, dataset_id='dataset1', resource_id=self.resource1['id'], user=None,
-            status=403
-        )
-
-    def test_resource_download_no_access(self, app):
-        resp = self.make_resource_download_request(
-            app, dataset_id='dataset1', resource_id=self.resource1['id'], user='user3',
-            status=403
-        )
-
-    def test_resource_download_collaborator(self, app):
-        core_helpers.call_action(
-            'package_collaborator_create',
-            id='dataset1',
-            user_id='user3',
-            capacity='member',
-        )
-        resp = self.make_resource_download_request(
-            app, dataset_id='dataset1', resource_id=self.resource1['id'], user='user3',
-            status=200
-        )
-
-    def test_resource_download_bad_resource(self, app):
-        resp = self.make_resource_download_request(
-            app, dataset_id='dataset1', resource_id='bad', user='user1',
-            status=404
-        )
-
-    def test_resource_download_valid(self, app):
-        sql = select([
-            model.Activity
-        ]).where(
-            and_(
-                model.Activity.activity_type == 'download resource',
-                model.Activity.object_id == self.dataset1['id'],
-                model.Activity.user_id == 'user1',
-            )
-        )
-
-        # before we start, this user has never downloaded this resource before
-        result = model.Session.execute(sql).fetchall()
-        assert 0 == len(result)
-
-        resp = self.make_resource_download_request(
-            app, dataset_id='dataset1', resource_id=self.resource1['id'], user='user1',
-            status=200
-        )
-
-        # after we've downloaded the resource, we should also
-        # have also logged a 'download resource' action for this user/resource
-        result = model.Session.execute(sql).fetchall()
-        assert 1 == len(result)
 
     # Publish Microdata
 
@@ -384,6 +222,7 @@ class TestExtendedPackageController(object):
             )
         mock_action.assert_called_once()
         assert 'failed for the following reason: &#34;oh no!&#34;' in resp.body
+
 
     # Request Access
